@@ -33,6 +33,8 @@
   "devDependencies": {
     "@babel/cli": "^7.15.7",
     "@babel/core": "^7.15.8",
+    "@babel/plugin-proposal-class-properties": "^7.18.6",
+    "@babel/plugin-proposal-object-rest-spread": "^7.20.7",
     "@babel/preset-env": "^7.15.8",
     "@codemirror/autocomplete": "^6.18.4",
     "@codemirror/commands": "^6.7.1",
@@ -48,23 +50,21 @@
     "@electron-forge/plugin-webpack": "^6.0.1",
     "@electron/notarize": "^3.0.1",
     "@electron/osx-sign": "^2.0.0",
-    "@johmun/vue-tags-input": "^1.5.10",
-    "@tiptap/extension-link": "^2.12.0",
-    "@tiptap/extension-underline": "^2.12.0",
-    "@tiptap/pm": "^2.12.0",
-    "@tiptap/starter-kit": "^2.12.0",
-    "@tiptap/vue-2": "^2.12.0",
+    "@sipec/vue3-tags-input": "^3.0.4",
+    "@tiptap/extension-link": "^3.2.0",
+    "@tiptap/extension-underline": "^3.2.0",
+    "@tiptap/starter-kit": "^3.2.0",
+    "@tiptap/vue-3": "^3.2.0",
     "@vercel/webpack-asset-relocator-loader": "^1.7.3",
-    "@vue/test-utils": "^1.0.0-beta.29",
     "axios": "^0.22.0",
     "babel-loader": "^8.2.2",
+    "bufferutil": "^4.0.9",
     "codemirror": "^6.0.1",
     "codemirror-lang-schemioscript": "^0.1.3",
     "copy-webpack-plugin": "^11.0.0",
     "css-loader": "^6.7.1",
     "electron": "21.2.3",
     "electron-icon-maker": "^0.0.5",
-    "esm": "^3.2.25",
     "expect": "^27.4.2",
     "file-loader": "^0.11.2",
     "gifenc": "^1.0.3",
@@ -79,12 +79,12 @@
     "shortid": "^2.2.16",
     "style-loader": "^3.3.1",
     "thememirror": "^2.0.1",
-    "vue": "^2.6.10",
-    "vue-gtag": "^1.16.1",
-    "vue-loader": "^15.9.8",
-    "vue-router": "^2.7.0",
-    "vue-template-compiler": "^2.6.10",
-    "vuex": "^3.5.1",
+    "tiny-emitter": "^2.1.0",
+    "utf-8-validate": "^5.0.10",
+    "vue": "^3.5.13",
+    "vue-loader": "^17.4.2",
+    "vue-router": "^4.5.0",
+    "vuex": "^4.1.0",
     "webpack": "^5.64.0",
     "webpack-cli": "^4.9.1",
     "webpack-dev-server": "^4.5.0"
