@@ -4,8 +4,8 @@
 
 // This is an entry point for schemio bundle that is used for FS based Schemio app
 
-import Vue from 'vue';
-import VueRouter from 'vue-router';
+import {createApp} from 'vue';
+import {createRouter, createWebHistory} from 'vue-router';
 import App from './app/App.vue';
 import FolderView from './app/view/FolderView.vue';
 import SearchView from './app/view/SearchView.vue';
@@ -13,7 +13,6 @@ import AboutView from './app/view/AboutView.vue';
 import store from './store/Store.js';
 import SchemeEditorWebView from './app/view/SchemeEditorWebView.vue';
 import NotFoundView from './app/view/NotFoundView.vue';
-import { applyVueFilters } from './vue.filters';
 import Header from './app/components/Header.vue';
 import Footer from './app/components/Footer.vue';
 import { fsClientProvider } from './app/client/fsClient';
@@ -24,13 +23,6 @@ import { generateMapPatch, generateSchemePatch } from './scheme/SchemePatch';
 import { processJSONTemplate } from './templater/templater.js';
 import { defaultStarterTemplates } from './components/editor/DefaultStarterTemplates.js';
 
-const sanitizeHtml = require('sanitize-html');
-
-Vue.use(VueRouter);
-applyVueFilters(Vue);
-
-Vue.component('schemio-header', Header);
-Vue.component('schemio-footer', Footer);
 
 const routePrefix = document.body.getAttribute('data-route-prefix') || '';
 store.dispatch('setRoutePrefix', routePrefix);
@@ -45,7 +37,6 @@ window.schemioDebug = {
     testAST,
     processJSONTemplate,
     generateMapPatch,
-    sanitizeHtml,
     testEvalAST : (expr, data) => testAST(expr).evalNode(new Scope(data))
 };
 
@@ -59,19 +50,20 @@ const routes = [
     route('NotFoundView',               `${routePrefix}/not-found`,        NotFoundView),
     route('HomeView',                   rootPath,                          FolderView, {clientProvider: fsClientProvider}),
     route('SearchView',                 `${routePrefix}/search`,           SearchView, {clientProvider: fsClientProvider}),
-    route('FolderView',                 `${routePrefix}/f/*`,              FolderView, {clientProvider: fsClientProvider}),
-    { path: '/', redirect: rootPath},
-    { path: '*', redirect: `${routePrefix}/not-found`}
+    route('FolderView',                 `${routePrefix}/f/:folders*`,      FolderView, {clientProvider: fsClientProvider}),
+    { path: '/.*', redirect: `${routePrefix}/not-found`}
 ];
 
 
-const router = new VueRouter({
-    mode: 'history',
+const router = createRouter({
+    history: createWebHistory(),
     routes: routes,
 });
 
 
-new Vue(Vue.util.extend({
-    router,
-    store,
-}, App)).$mount('#app');
+const app = createApp(App);
+app.use(router);
+app.use(store);
+app.component('schemio-header', Header);
+app.component('schemio-footer', Footer);
+app.mount('#app');
