@@ -1,9 +1,12 @@
-import Vue from 'vue';
+// const Emitter = require('tiny-emitter');
+
+import Emitter from 'tiny-emitter';
+
 import {Logger} from '../../logger';
 
 const log = new Logger('EditorEventBus');
 
-const bus = new Vue({});
+const bus = new Emitter();
 
 function generateEvent(editorId, eventName, args) {
     let fullEvent = editorId + '/' + eventName;
@@ -17,13 +20,13 @@ function $on(editorId, eventName, args, callback) {
     if (!editorId) {
         throw new Error('editorId should be specified, got: ', editorId);
     }
-    bus.$on(generateEvent(editorId, eventName, args), callback);
+    bus.on(generateEvent(editorId, eventName, args), callback);
 }
 function $off(editorId, eventName, args, callback) {
     if (!editorId) {
         throw new Error('editorId should be specified, got: ', editorId);
     }
-    bus.$off(generateEvent(editorId, eventName, args), callback);
+    bus.off(generateEvent(editorId, eventName, args), callback);
 }
 function $emit(editorId, eventName, eventArgs, ...emitArgs) {
     if (!editorId) {
@@ -31,7 +34,7 @@ function $emit(editorId, eventName, eventArgs, ...emitArgs) {
     }
     const fullEventName = generateEvent(editorId, eventName, eventArgs);
     log.infoEvent(fullEventName, emitArgs);
-    bus.$emit(fullEventName, ...emitArgs);
+    bus.emit(fullEventName, ...emitArgs);
 }
 
 const EditorEventBus = {
@@ -154,16 +157,16 @@ const EditorEventBus = {
             any: {
                 $on: (editorId, callback) => $on(editorId, 'any-text-slot-triggered', [], callback),
                 $off: (editorId, callback) => $off(editorId, 'any-text-slot-triggered', [], callback),
-                $emit: (editorId, item, textSlot, creatingNewItem) => {
-                    $emit(editorId, 'any-text-slot-triggered', [], item, textSlot, creatingNewItem);
+                $emit: (editorId, item, slotName, area, markupDisabled, creatingNewItem) => {
+                    $emit(editorId, 'any-text-slot-triggered', [], item, slotName, area, markupDisabled, creatingNewItem);
                 }
             },
             specific: {
                 $on: (editorId, itemId, callback) => $on(editorId, 'text-slot-triggered', [itemId], callback),
                 $off: (editorId, itemId, callback) => $off(editorId, 'text-slot-triggered', [itemId], callback),
-                $emit: (editorId, item, textSlot, creatingNewItem) => {
-                    EditorEventBus.textSlot.triggered.any.$emit(editorId, item, textSlot, creatingNewItem);
-                    $emit(editorId, 'text-slot-triggered', [item.id], item, textSlot, creatingNewItem);
+                $emit: (editorId, item, slotName, area, markupDisabled, creatingNewItem) => {
+                    EditorEventBus.textSlot.triggered.any.$emit(editorId, item, slotName, area, markupDisabled, creatingNewItem);
+                    $emit(editorId, 'text-slot-triggered', [item.id], item, slotName, area, markupDisabled, creatingNewItem);
                 }
             },
         },
@@ -189,6 +192,16 @@ const EditorEventBus = {
             $off: (editorId, callback) => $off(editorId, 'text-slot-moved', [], callback),
             $emit: (editorId, item, slotName, destinationSlotName) => {
                 $emit(editorId, 'text-slot-moved', [], item, slotName, destinationSlotName);
+            }
+        }
+    },
+
+    inPlaceTextEditor: {
+        created: {
+            $on: (editorId, callback) => $on(editorId, 'in-place-text-editor-created', [], callback),
+            $off: (editorId, callback) => $off(editorId, 'in-place-text-editor-created', [], callback),
+            $emit: (editorId, editor) => {
+                $emit(editorId, 'in-place-text-editor-created', [], editor);
             }
         }
     },
