@@ -105,26 +105,28 @@
                 </div>
                 <table v-else class="properties-table">
                     <tbody>
-                        <tr v-for="(arg, argName) in shapeComponent.args" v-if="shapePropsControlStates[argName] && !isArgumentHidden(arg)">
-                            <td class="label" width="50%" :class="{disabled: !shapePropsControlStates[argName].shown}">
-                                {{arg.name}}
-                                <tooltip v-if="arg.description" :disabled="!shapePropsControlStates[argName].shown">{{arg.description}}</tooltip>
-                            </td>
-                            <td class="value" width="50%">
-                                <PropertyInput
-                                    :key="`prop-input-${item.id}-${item.shape}-${argName}-${item.shapeProps.thick}`"
-                                    :editorId="editorId"
-                                    :descriptor="arg"
-                                    :value="item.shapeProps[argName]"
-                                    :shapeProps="item.shapeProps"
-                                    :disabled="!shapePropsControlStates[argName].shown"
-                                    :schemeContainer="schemeContainer"
-                                    :leftOriented="argName === 'sourceCap'"
-                                    :itemId="item.id"
-                                    @input="onShapePropChange(argName, arg.type, $event)"
-                                />
-                            </td>
-                        </tr>
+                        <template v-for="(arg, argName) in shapeComponent.args">
+                            <tr v-if="shapePropsControlStates[argName] && !isArgumentHidden(arg)">
+                                <td class="label" width="50%" :class="{disabled: !shapePropsControlStates[argName].shown}">
+                                    {{arg.name}}
+                                    <tooltip v-if="arg.description" :disabled="!shapePropsControlStates[argName].shown">{{arg.description}}</tooltip>
+                                </td>
+                                <td class="value" width="50%">
+                                    <PropertyInput
+                                        :key="`prop-input-${item.id}-${item.shape}-${argName}-${item.shapeProps.thick}`"
+                                        :editorId="editorId"
+                                        :descriptor="arg"
+                                        :value="item.shapeProps[argName]"
+                                        :shapeProps="item.shapeProps"
+                                        :disabled="!shapePropsControlStates[argName].shown"
+                                        :schemeContainer="schemeContainer"
+                                        :leftOriented="argName === 'sourceCap'"
+                                        :itemId="item.id"
+                                        @input="onShapePropChange(argName, arg.type, $event)"
+                                    />
+                                </td>
+                            </tr>
+                        </template>
                     </tbody>
                 </table>
             </panel>
