@@ -1,21 +1,17 @@
 /* This Source Code Form is subject to the terms of the Mozilla Public
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at https://mozilla.org/MPL/2.0/. */
-import Vue from 'vue';
-import Vuex from 'vuex';
+import {createStore} from 'vuex';
 import {forEach} from '../collections';
 import utils from '../utils';
 import {createSettingStorageFromLocalStorage} from '../LimitedSettingsStorage';
 import shortid from 'shortid';
 
-Vue.use(Vuex);
-
 const DEFAULT_CONNECTOR_SMOOTHING = 'defaultConnectorSmoothing';
 
 const myStorage = createSettingStorageFromLocalStorage('store', 100);
 
-
-const store = new Vuex.Store({
+const store = createStore({
     state: {
         //rootPath is used in the header for a home link, since if this is being hosted in GitHub Pages then we cannot use '/' as root path
         rootPath: '/',
@@ -88,7 +84,6 @@ const store = new Vuex.Store({
             connectorItemId: null,
             mx: 0, // value on x viewport axis
             my: 0, // value on y viewport axis
-            primaryShapeId: null,
         },
 
         // used for storing information about images that were dropped on svg editor
@@ -192,12 +187,11 @@ const store = new Vuex.Store({
             state.snappers.vertical = null;
         },
 
-        PROPOSE_CONNECTOR_DESTINATION_ITEMS(state, {connectorItemId, mx, my, primaryShapeId}) {
+        PROPOSE_CONNECTOR_DESTINATION_ITEMS(state, {connectorItemId, mx, my}) {
             state.connectorProposedDestination.connectorItemId = connectorItemId;
             state.connectorProposedDestination.mx = mx;
             state.connectorProposedDestination.my = my;
             state.connectorProposedDestination.shown = true;
-            state.connectorProposedDestination.primaryShapeId = primaryShapeId;
         },
         DISABLE_PROPOSE_CONNECTOR_DESTINATION_ITEMS(state) {
             state.connectorProposedDestination.shown = false;
@@ -449,8 +443,8 @@ const store = new Vuex.Store({
             commit('TOGGLE_ITEM_DETAIL_MARKERS');
         },
 
-        proposeConnectorDestinationItems({commit}, {connectorItemId, mx, my, primaryShapeId}) {
-            commit('PROPOSE_CONNECTOR_DESTINATION_ITEMS', {connectorItemId, mx, my, primaryShapeId});
+        proposeConnectorDestinationItems({commit}, payload) {
+            commit('PROPOSE_CONNECTOR_DESTINATION_ITEMS', payload);
         },
         disableProposeConnectorDestinationItems({commit}) {
             commit('DISABLE_PROPOSE_CONNECTOR_DESTINATION_ITEMS');
